requirements.txt
appdirs==1.4.3
Django==1.11.1
lxml==3.7.3
packaging==16.8
pyparsing==2.2.0
python-libtorrent==1.1.3
pytz==2017.2
requests==2.14.2
six==1.10.0
